% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyEdu-package.r
\docType{package}
\name{eyEdu}
\alias{eyEdu}
\alias{eyEdu-package}
\title{eyEdu.}
\description{
EyEdu stands for eye-movement analysis for educational purposes.
The package is NOT intended to replace any proper analysis software for eye-tracking data. It
is rather intended to provide some basic functions that may help students who make their first
steps into eye-tracking to get a better understanding of the method. Example experiments and data
recorded using a low-cost eye-tracker (EyeTribe) are provided.
}
